refactor(cpp): split into cpp library and Python bindings
- Remove Python bindings (Pybind11 interface) from cpp library
- Update include directives to use new header structure
- Add build script for automated compilation
- Moved global variables section to headers
- Create separate Python bindings library (acs_int)
- Added devcontainer (regular ubuntu noble)
