+++ .devcontainer/post_create.sh
@@ -1,0 +1,7 @@
+#!/bin/bash
+
+# Activate .venv environment
+source ../.venv/bin/activate
+
+# Install requirements
+pip install -r requirements.txt
